+++ sidex/src/sidexInteger.cpp
@@ -110,7 +110,7 @@ char* sidexInteger::toString()
   return (char*)sReturnString.data();
 #else // USESTREAMS
 #if defined(LINUX) || defined (MINGW_BUILD)
-  sprintf(m_strRepresentation, "%lld", value);
+  sprintf(m_strRepresentation, "%lld", (long long) value);
 #else // LINUX
   sprintf_s(m_strRepresentation, 64, "%lld", value);
 #endif // LINUX
